Revert "Update EnergyData.cs"
This reverts commit a19c3a0a1c893674cbba4e35caec8f21357f0c44.

+++ Heatington/Optimizer/EnergyData.cs
@@ -8,10 +8,10 @@ public struct EnergyData
     [Index(1)]
     public DateTime EndTime { get; set; }
     [Index(2)]
-    public double HeatDemandMwh { get; set; }
+    public float HeatDemandMwh { get; set; }
     [Index(3)]
-    public double ElectricityPrice { get; set; }
-
+    public float ElectricityPrice { get; set; }
+    
     public override string ToString()
     {
         string s = string.Concat(StartTime," ", EndTime," ",HeatDemandMwh," ",ElectricityPrice);
